domain/user/user_router.py
from datetime import timedelta, datetime, UTC

from fastapi import APIRouter, HTTPException, Depends
from fastapi.security import OAuth2PasswordRequestForm, OAuth2PasswordBearer
from jose import jwt, JWTError
from sqlalchemy.orm import Session
from starlette import status

from database import get_db
from domain.user import user_crud, user_schema
from domain.user.user_crud import pwd_context
from domain.user.user_schema import PasswordVerify

ACCESS_TOKEN_EXPIRE_MINUTES = 60 * 24
# $ openssl rand -hex 32
SECRET_KEY = "d636517809afb1541819a0414b13574f329b4266a49b35246478f4017269fbf8"
ALGORITHM = "HS256"
oauth2_scheme = OAuth2PasswordBearer(tokenUrl="/api/user/login")


router = APIRouter(
    prefix="/api/user",
)


@router.post("/create", status_code=status.HTTP_204_NO_CONTENT)
def user_create(
        _user_create: user_schema.UserCreate,
        db: Session = Depends(get_db)
):
    user = user_crud.get_existing_user(db, user_create=_user_create)
    if user:
        raise HTTPException(status_code=status.HTTP_409_CONFLICT,
                            detail="이미 존재하는 사용자입니다.")
    user_crud.create_user(db=db, user_create=_user_create)


@router.post("/login", response_model=user_schema.Token)
def login_for_access_token(
        form_data: OAuth2PasswordRequestForm = Depends(),
       db: Session = Depends(get_db)
):
    # 아이디 확인
    user = user_crud.get_user(db, form_data.username)
    if not user:
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail="가입되지 않은 아이디입니다."
        )

    # 비밀번호 확인
    if not user or not pwd_context.verify( form_data.password, user.password ):
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="비밀번호가 일치하지 않습니다.",
            headers={"WWW-Authenticate": "Bearer"},
        )

    # make access token
    data = {
        "sub": user.username,
        "exp": datetime.now(UTC) + timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
    }
    access_token = jwt.encode(data, SECRET_KEY, algorithm=ALGORITHM)

    return {
        "access_token": access_token,
        "token_type": "bearer",
        "username": user.username
    }

def get_current_user(
        token: str = Depends(oauth2_scheme),
        db: Session = Depends(get_db)
):
    credentials_exception = HTTPException(
        status_code=status.HTTP_401_UNAUTHORIZED,
        detail="Could not validate credentials",
        headers={"WWW-Authenticate": "Bearer"},
    )
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        username: str = payload.get("sub")
        if username is None:
            raise credentials_exception
    except JWTError:
        raise credentials_exception
    else:
        user = user_crud.get_user(db, username=username)
        if user is None:
            raise credentials_exception
        return user

@router.post("/verify-password", status_code=204)
def verify_password(
    data: PasswordVerify,
    db: Session = Depends(get_db),
    current_user: user_schema.User = Depends(get_current_user)
):
    if not user_crud.verify_user_password(db, current_user.username, data.password):
        raise HTTPException(
            status_code=401,
            detail="비밀번호가 일치하지 않습니다."
        )

@router.delete("/delete", status_code=204)
def delete_current_user(
    db: Session = Depends(get_db),
    current_user: user_schema.User = Depends(get_current_user)
):
    user_crud.delete_user(db, current_user.username)
