requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.0
fastapi==0.115.12
greenlet==3.2.2
h11==0.16.0
httptools==0.6.4
idna==3.10
pydantic==2.11.4
pydantic-core==2.33.2
python-dotenv==1.1.0
pyyaml==6.0.2
sniffio==1.3.1
sqlalchemy==2.0.40
starlette==0.46.2
typing-extensions==4.13.2
typing-inspection==0.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1

fastapi
uvicorn
sqlalchemy
pydantic
passlib[bcrypt]
python-jose
pydantic[email]
python-multipart
